Refactored the step to be an operator instead of a large file in the runner package.
