+++ src/runner/__init__.py
@@ -11,8 +11,8 @@ Public API::
 
 from .run import init_state
 from .run import run
-from .step import step_multiphase
-from .step import step_single_phase
+from operators.step import step_multiphase
+from operators.step import step_single_phase
 
 __all__ = [
     "init_state",
